+++ app/forms/search_form.rb
@@ -3,6 +3,7 @@ class SearchForm
   include ActiveModel::Validations::Callbacks
 
   include ::Filters::PensionPot
+  include ::Filters::QualificationOrAccreditation
   include ::Filters::Language
 
   ADVICE_METHOD_FACE_TO_FACE = 'face_to_face'
@@ -27,7 +28,6 @@ class SearchForm
                 :postcode,
                 :coordinates,
                 :firm_id,
-                :qualification_or_accreditation,
                 :random_search_seed,
                 *TYPES_OF_ADVICE,
                 *OTHER_SERVICES,
@@ -49,10 +49,6 @@ class SearchForm
 
   def coordinates
     @coordinates ||= Geocode.call(postcode)
-  end
-
-  def options_for_qualifications_and_accreditations
-    (options_for(Qualification) + options_for(Accreditation)).sort
   end
 
   def retirement_income_products?
@@ -81,14 +77,6 @@ class SearchForm
     phone_or_online? ? OtherAdviceMethod.all.map(&:id) : []
   end
 
-  def selected_qualification_id
-    selected_filter_id_for(Qualification)
-  end
-
-  def selected_accreditation_id
-    selected_filter_id_for(Accreditation)
-  end
-
   def postcode
     @postcode if face_to_face?
   end
@@ -108,37 +96,7 @@ class SearchForm
     postcode.try(:upcase!)
   end
 
-  def prefix_for(model)
-    model.model_name.singular[0]
-  end
-
-  def filters_for(model)
-    key_to_i = -> (k, v) { [k.to_s.to_i, v] }
-    I18n.t("search.filter.#{model.model_name.i18n_key}.ordinal").map(&key_to_i).to_h
-  end
-
-  def options_for(model)
-    filters = filters_for(model)
-    model
-      .where(order: filters.keys)
-      .pluck(:order, :id)
-      .map { |order, id| [filters[order], "#{prefix_for(model)}#{id}"] }
-  end
-
   def selected_checkbox_attributes(attribute_names)
     attribute_names.select { |type| public_send(type) == '1' }
-  end
-
-  def selected_filter_id_for(model)
-    is_desired_type = ->(prefix, item) { !!item[/^#{prefix}/] }
-    extract_id = ->(item) { item[1..-1] }
-    type_prefix = prefix_for(model)
-
-    [qualification_or_accreditation]
-      .compact
-      .select(&is_desired_type.curry[type_prefix])
-      .map(&extract_id)
-      .map(&:to_i)
-      .first
   end
 end
